requirements.txt
lxml==4.5.2
numpy==1.18.5
pillow
torch>=1.5.0
torchvision>=0.6.0
opencv-contrib-python
kornia>=0.3.1